# internal/models/model_data_point.go
/*
 * Senso-Care
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package models

type DataPoint struct {

	Date string `json:"date,omitempty"`

	Value float32 `json:"value,omitempty"`

	Info string `json:"info,omitempty"`
}