# internal/models/model_metric.go
/*
 * Senso-Care
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package models

type Metric struct {

	Type string `json:"type,omitempty"`

	Sensors []SensorData `json:"sensors,omitempty"`
}