# internal/models/model_sensor_data.go
/*
 * Senso-Care
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package models

type SensorData struct {

	Name string `json:"name,omitempty"`

	Series []DataPoint `json:"series,omitempty"`
}